Add cell app table
